refactor: move model utils to ml package (can later become a class)
