+++ src/safeds/ml/classification/_random_forest.py
@@ -1,7 +1,7 @@
 from typing import Optional
 
 # noinspection PyProtectedMember
-import safeds._util._util_sklearn
+import safeds.ml._util_sklearn
 from safeds.data import SupervisedDataset
 from safeds.data.tabular import Table
 from sklearn.ensemble import RandomForestClassifier
@@ -31,7 +31,7 @@ class RandomForest:
         LearningError
             If the supervised dataset contains invalid values or if the training failed.
         """
-        self.target_name = safeds._util._util_sklearn.fit(
+        self.target_name = safeds.ml._util_sklearn.fit(
             self._classification, supervised_dataset
         )
 
@@ -56,7 +56,7 @@ class RandomForest:
         PredictionError
             If prediction with the given dataset failed.
         """
-        return safeds._util._util_sklearn.predict(
+        return safeds.ml._util_sklearn.predict(
             self._classification,
             dataset,
             target_name if target_name is not None else self.target_name,
